# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ClusterRenameRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
    type: object
  dto.ClusterResponse:
    properties:
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      lookup_key:
        type: string
      member_count:
        type: integer
    type: object
  dto.DiscoveryJobRequest:
    properties:
      snmp_community:
        type: string
      snmp_version:
        type: string
      targets:
        type: string
    type: object
  dto.DiscoveryJobResponse:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      error_message:
        type: string
      id:
        type: string
      input_spec:
        type: string
      snmp_version:
        type: string
      started_at:
        type: string
      status:
        type: string
      total_targets:
        type: integer
    type: object
  dto.DiscoveryJobResultResponse:
    properties:
      element_id:
        type: string
      error_message:
        type: string
      ip_address:
        type: string
      responded_at:
        type: string
      status:
        type: string
      sys_descr:
        type: string
      sys_name:
        type: string
      sys_object_id:
        type: string
    type: object
...
info:
  contact: {}
  description: SNMP discovery + rule-based virtual cluster assignment for network
    devices (Phase 1).
  title: Netcluster API
  version: "1.0"
paths:
...
  /clusters/{id}:
    delete:
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: cluster still has member elements
          schema:
            additionalProperties: true
            type: object
      summary: Delete an empty cluster
      tags:
      - clusters
    get:
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClusterResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a cluster by ID
      tags:
      - clusters
    patch:
      consumes:
      - application/json
      description: Renaming never touches lookup_key, so the rule engine keeps re-finding
        this cluster on future matches.
      parameters:
      - description: Cluster ID
        in: path
        name: id
        required: true
        type: string
      - description: New display name/description
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ClusterRenameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ClusterResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Rename a cluster's display name/description
      tags:
      - clusters
  /discovery/jobs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DiscoveryJobResponse'
            type: array
      summary: List discovery jobs
      tags:
      - discovery
    post:
      consumes:
      - application/json
      parameters:
      - description: IP pool / CIDR / range plus SNMP settings
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DiscoveryJobRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dto.DiscoveryJobResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Start an async SNMP discovery scan
      tags:
      - discovery
  /discovery/jobs/{jobId}:
    get:
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DiscoveryJobResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get a discovery job's status
      tags:
      - discovery
  /discovery/jobs/{jobId}/cancel:
    post:
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Best-effort cancel a running discovery job
      tags:
      - discovery
  /discovery/jobs/{jobId}/results:
    get:
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      - description: Max rows to return (default 100, max 1000)
        in: query
        name: limit
        type: integer
      - description: Rows to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DiscoveryJobResultResponse'
            type: array
      summary: List a discovery job's per-IP results
      tags:
      - discovery
...
swagger: "2.0"